Extract mouse UI simulation activity logging
Move the structured start and completion log construction into a cohesive helper while preserving every event name, message, context field, and correlation identifier.
