+++ Packages/src/Editor/FirstPartyTools/SimulateMouseUi/SimulateMouseUiUseCase.cs
@@ -64,7 +64,7 @@ namespace io.github.hatayama.UnityCliLoop.FirstPartyTools
             Debug.Assert(eventSystem != null, "ValidateSimulationStart must reject a missing EventSystem.");
             EventSystem activeEventSystem = eventSystem!;
 
-            LogSimulationStart(parameters, correlationId);
+            MouseUiSimulationActivityLogger.LogSimulationStart(parameters, correlationId);
             EnsureOverlayExists();
 
             SimulateMouseUiResponse? dragStateFailure =
@@ -76,40 +76,9 @@ namespace io.github.hatayama.UnityCliLoop.FirstPartyTools
 
             SimulateMouseUiResponse response =
                 await ExecuteMouseAction(parameters, activeEventSystem, ct).ConfigureAwait(false);
-            LogSimulationComplete(parameters, response, correlationId);
+            MouseUiSimulationActivityLogger.LogSimulationComplete(parameters, response, correlationId);
 
             return response;
-        }
-
-        internal static void LogSimulationStart(MouseUiSimulationCommand parameters, string correlationId)
-        {
-            VibeLogger.LogInfo(
-                "simulate_mouse_start",
-                "Mouse simulation started",
-                new
-                {
-                    Action = parameters.Action.ToString(),
-                    X = parameters.X,
-                    Y = parameters.Y,
-                    BypassRaycast = parameters.BypassRaycast,
-                    TargetPath = parameters.TargetPath,
-                    DropTargetPath = parameters.DropTargetPath
-                },
-                correlationId: correlationId
-            );
-        }
-
-        internal static void LogSimulationComplete(
-            MouseUiSimulationCommand parameters,
-            SimulateMouseUiResponse response,
-            string correlationId)
-        {
-            VibeLogger.LogInfo(
-                "simulate_mouse_complete",
-                $"Mouse simulation completed: {response.Message}",
-                new { Action = parameters.Action.ToString(), Success = response.Success },
-                correlationId: correlationId
-            );
         }
 
         private async Task<SimulateMouseUiResponse> ExecuteMouseAction(
